Cast distanceFromLocationToRestaurant to int.

+++ app/src/main/java/com/example/go4lunch/model/Restaurant.java
@@ -51,7 +51,7 @@ public class Restaurant implements Parcelable {
         }
     };
 
-    public float getDistanceFromDeviceLocation() {
+    public int getDistanceFromDeviceLocation() {
         return distanceFromDeviceLocation;
     }
